Changes in tests and tests framework for testing with SkySQL
Introduced new option (-a or ENV{TEST_ADD_PARAM}) for supplying
additional connection string parameters. That may be required for tests
opening additional connections not using the TEST_DSN. Good example
would be required encrypted connection parameters.
Some tests have been fixed to be more durable for different setups.

+++ test/relative.c
@@ -682,7 +682,7 @@ ODBC_TEST(t_rows_fetched_ptr1)
   rc = SQLFetchScroll(Stmt,SQL_FETCH_NEXT,0);
   while (rc == SQL_SUCCESS || rc == SQL_SUCCESS_WITH_INFO)
   {
-    fprintf(stdout,"total rows fetched: %ld\n", rowsFetched);
+    fprintf(stdout,"total rows fetched: %ld\n", (long)rowsFetched);
     IS(rowsFetched == rowsSize);
     i++; rowsFetched= 0;
     rc = SQLFetchScroll(Stmt, SQL_FETCH_NEXT, 0);
@@ -704,7 +704,7 @@ ODBC_TEST(t_rows_fetched_ptr1)
   rc = SQLFetchScroll(Stmt,SQL_FETCH_NEXT,0);
   while (rc == SQL_SUCCESS || rc == SQL_SUCCESS_WITH_INFO)
   {
-    fprintf(stdout,"total rows fetched: %ld\n", rowsFetched);
+    fprintf(stdout,"total rows fetched: %ld\n", (long)rowsFetched);
     IS(rowsFetched == rowsSize);
     i++;rowsFetched= 0;
     rc = SQLFetchScroll(Stmt,SQL_FETCH_NEXT,0);
